refactor: rename input component
